+++ CHANGES.md
@@ -6,7 +6,10 @@ This list does not include all the tiny changes, bug handling, etc, only the cha
 
 - If applicable, the reference to the generated context file is, if requested via an appropriate ID, added to the HTML content.
 - [Bug, sort of] Using `@import` in the generated json-ld context turned out to be brittle. Changed the context to an array, with the generated portion preceded by the list of "imported" context files (if applicable).
-- Introduced the artificial range `langString` to denote a property whose values are natural language strings. The main issue is that these properties should only have an `id` value defined in the generated context file, otherwise type coercion will remove the effects of language and direction setting. An English remark is generated in the HTML text.
+- Special actions when the `rdf:langString` is part of the range:
+    - the `rdf:dirLangString` is added to the range as a future-proof feature eyeing towards RDF 1.2
+    - if the `range_union` flag is not set to `true` an error is raised, because that is the only way this can work in practice
+    - in the generated context the property should only have an `id`, even if it is combined with `xsd:string`; otherwise the language/directions tags will be ignored
 
 ## Version 1.8.1